Fixed up remote IDs
* Namespaced remote persistence UUIDs with the remote universe, and made them parsable as hook parts
* Parsed hook parts and pointed them at the actual remote UUIDs they match
